Support more cross targets
Adds support for Android, Linux/ARM, and windows. Bare-metal thumb targets are listed,
because they're supported by cross, but they aren't enabled by default as they don't
support std and will likely not be used by most projects.

I've left off the *linux-musleabi* targets as they are failing with an odd error that
is likely due to musl itself.

Additionally I've left out iOS targets as that support needs to be added to trust before
it should be added here.